pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "virtual-accelerator"
dynamic = ["version"]
description = "SLAC Specific implementation of virtual accelerators for SLAC beamlines"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "SLAC National Accelerator Laboratory"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pandas",
    "numpy",
    "scipy",
    "pyyaml",
    "lume-base @ git+https://github.com/slaclab/lume-base",
    "pmd-beamphysics",
]

[project.urls]
Homepage = "https://github.com/slaclab/virtual-accelerator"
Repository = "https://github.com/slaclab/virtual-accelerator"
Documentation = "https://github.com/slaclab/virtual-accelerator"

[project.optional-dependencies]
bmad = [
    "lume-bmad @ git+https://github.com/lume-science/lume-bmad",
]
cheetah = [
    "lume-cheetah @ git+https://github.com/lume-science/lume-cheetah"
]
impact = [
    "lume-impact @ git+https://github.com/ChristopherMayes/lume-impact"
]
pva = [
    "lume-pva @ git+https://github.com/lume-science/lume-pva"
]
surrogate = [
    "lume-torch @ git+https://github.com/lume-science/lume-torch"
]
all = [
    "lume-bmad @ git+https://github.com/lume-science/lume-bmad",
    "lume-cheetah @ git+https://github.com/lume-science/lume-cheetah",
    "lume-impact @ git+https://github.com/ChristopherMayes/lume-impact",
    "lume-torch @ git+https://github.com/lume-science/lume-torch",
    "lume-pva @ git+https://github.com/lume-science/lume-pva",
    "pmd-beamphysics"
]
dev = [
    "pytest",
    "ipykernel",
    "pre-commit"
]
docs = [
    "pygments",
    "mkdocs",
    "mkdocstrings",
    "mkdocstrings-python",
    "mkdocs-material",
    "livereload",
    "pytkdocs[numpy-style]",
]

[tool.setuptools.packages.find]
include = [ "virtual_accelerator", ]
namespaces = false

[tool.setuptools.package-data]
virtual_accelerator = ["py.typed"]

[tool.setuptools_scm]
version_file = "virtual_accelerator/_version.py"

[tool.ruff]
# select = []
# ignore = []
